add hero component
